fix: move app pages to [locale] segment and use relative import for translations
- Use relative path in i18n/request.ts to load translations (fixes dev module error)
- Move pages under src/app/[locale]/ to support next-intl prefix routing
- Keep root layout.tsx minimal; localized layout in [locale]/layout.tsx
