+++ src/app/[locale]/privacy-policy/page.tsx
