+++ src/app/[locale]/result/page.tsx
